refactor(plugin): simplify impl
