+++ src/plugin/factory.ts
@@ -1,17 +1,16 @@
-import { BasePlugin } from './base';
 import { topologicalSort } from './common';
-import { ArtusPlugin } from './impl';
-import { PluginConfigItem } from './types';
+import { Plugin } from './impl';
+import { PluginConfigItem, PluginMap, PluginType } from './types';
 
 export class PluginFactory {
-  static async create(name: string, item: PluginConfigItem): Promise<BasePlugin> {
-    const pluginInstance = new ArtusPlugin(name, item);
+  static async create(name: string, item: PluginConfigItem): Promise<PluginType> {
+    const pluginInstance = new Plugin(name, item);
     await pluginInstance.init();
     return pluginInstance;
   }
 
-  static async createFromConfig(config: Record<string, PluginConfigItem>): Promise<BasePlugin[]> {
-    const pluginInstanceMap: Map<string, BasePlugin> = new Map();
+  static async createFromConfig(config: Record<string, PluginConfigItem>): Promise<PluginType[]> {
+    const pluginInstanceMap: PluginMap = new Map();
     for (const [name, item] of Object.entries(config)) {
       const pluginInstance = await PluginFactory.create(name, item);
       if (pluginInstance.enable) {
@@ -32,9 +31,9 @@ export class PluginFactory {
     return pluginSortResult.map(name => pluginInstanceMap.get(name)!);
   }
 
-  static filterDuplicatePlugins(plugins: BasePlugin[]): BasePlugin[] {
+  static filterDuplicatePlugins(plugins: PluginType[]): PluginType[] {
     const exists: Map<string, boolean> = new Map();
-    const filtedPlugins: BasePlugin[] = [];
+    const filtedPlugins: PluginType[] = [];
     for (const plugin of plugins) {
       const key = plugin.importPath;
       if (exists.get(key)) {
