+++ src/plugin/impl.ts
@@ -1,10 +1,35 @@
 import path from 'path';
-import { BasePlugin } from './base';
 import { loadMetaFile } from '../utils/load_meta_file';
 import { exisis } from '../utils/fs';
 import { PLUGIN_META_FILENAME } from '../constant';
+import { PluginConfigItem, PluginMap, PluginMetadata, PluginType } from './types';
+import { getPackagePath } from './common';
 
-export class ArtusPlugin extends BasePlugin {
+export class Plugin implements PluginType {
+  public name: string;
+  public enable: boolean;
+  public importPath = '';
+  public metadata: Partial<PluginMetadata> = {};
+  public metaFilePath = '';
+
+  constructor(name: string, configItem: PluginConfigItem) {
+    this.name = name;
+    this.enable = configItem.enable ?? false;
+    if (this.enable) {
+      let importPath = configItem.path ?? '';
+      if (configItem.package) {
+        if (importPath) {
+          throw new Error(`plugin ${name} config error, package and path can't be set at the same time.`);
+        }
+        importPath = getPackagePath(configItem.package);
+      }
+      if (!importPath) {
+        throw new Error(`Plugin ${name} need have path or package field`);
+      }
+      this.importPath = importPath;
+    }
+  }
+
   async init() {
     if (!this.enable) {
       return;
@@ -16,6 +41,26 @@ export class ArtusPlugin extends BasePlugin {
     if (this.metadata.name !== this.name) {
       throw new Error(`${this.name} metadata invalid, name is ${this.metadata.name}`);
     }
+  }
+
+  public checkDepExisted(pluginMap: PluginMap) {
+    for (const { name: pluginName, optional } of this.metadata.dependencies ?? []) {
+      const instance = pluginMap.get(pluginName);
+      if (!instance || !instance.enable) {
+        if (optional) {
+          // TODO: use artus logger instead
+          console.warn(`Plugin ${this.name} need have optional dependence: ${pluginName}.`);
+        } else {
+          throw new Error(`Plugin ${this.name} need have dependence: ${pluginName}.`);
+        }
+      }
+    }
+  }
+
+  public getDepEdgeList(): [string, string][] {
+    return this.metadata.dependencies
+      ?.filter(({ optional }) => !optional)
+      ?.map(({ name: depPluginName }) => [this.name, depPluginName]) ?? [];
   }
 
   private async checkAndLoadMetadata() {
