+++ src/plugin/index.ts
@@ -1,5 +1,3 @@
-import {} from './types';
-
-export * from './base';
+export * from './types';
 export * from './impl';
 export * from './factory';
