+++ test/plugin.test.ts
@@ -1,6 +1,6 @@
 import 'reflect-metadata';
 import path from 'path';
-import { ArtusPlugin, PluginFactory } from '../src';
+import { Plugin, PluginFactory } from '../src';
 
 const pluginPrefix = 'fixtures/plugins';
 
@@ -45,7 +45,7 @@ describe('test/app.test.ts', () => {
       const pluginList = await PluginFactory.createFromConfig(mockPluginConfig);
       expect(pluginList.length).toEqual(3);
       pluginList.forEach(plugin => {
-        expect(plugin).toBeInstanceOf(ArtusPlugin);
+        expect(plugin).toBeInstanceOf(Plugin);
         expect(plugin.enable).toBeTruthy();
       });
       expect(pluginList.map(plugin => plugin.name)).toStrictEqual(['plugin-c', 'plugin-b', 'plugin-a']);
@@ -196,7 +196,7 @@ describe('test/app.test.ts', () => {
       const pluginList = await PluginFactory.createFromConfig(mockPluginConfig);
       expect(pluginList.length).toEqual(1);
       pluginList.forEach(plugin => {
-        expect(plugin).toBeInstanceOf(ArtusPlugin);
+        expect(plugin).toBeInstanceOf(Plugin);
         expect(plugin.enable).toBeTruthy();
       });
       expect(mockWarnFn).toBeCalledWith(`Plugin plugin-d need have optional dependence: plugin-e.`);
